Update script.js
Step 76

What if the player doesn't have enough gold to buy health? When you want to run code conditionally, you can use the if statement.

An if statement is used to make decisions in code. The keyword if is followed by a condition in parentheses. If the condition is true, the code inside the curly braces {} is executed. If the condition is false, the code inside the curly braces is skipped.

Here is an example of an if statement:

const num = 5;
if (num >= 3) {
  console.log("This code will run because num is greater than or equal to 3.");
}

Start by placing all of the code in your buyHealth function inside an if statement. For the if statement condition, check if gold is greater than or equal to 10.