Update script.js
Step 89

You can insert variables into a string with the concatenation operator +. Update the "You now have a new weapon." string to say "You now have a " and the name of the new weapon. Remember to end the sentence with a period.

Here is an example that creates the string "Hello, our name is freeCodeCamp.":

const ourName = "freeCodeCamp";
const ourStr = "Hello, our name is " + ourName + ".";