Update script.js
Step 138

In order for the &#x2620; emoticon text to properly display on the page, you will need to use the innerHTML property.

The innerHTML property allows you to access or modify the content inside an HTML element using JavaScript.

Here is an example of updating the content for this paragraph element using the innerHTML property.

<p id="demo">This is a paragraph.</p>

document.querySelector("#demo").innerHTML = "Hello, innerHTML!";

In the update function, change text.innerText to text.innerHTML.