Update script.js
Step 154

Remember that the increment operator ++ can be used to increase a variable's value by 1. There is also a decrement operator -- that can be used to decrease a variable's value by 1. For example :

let num = 10;
num--;
console.log(num); // Output: 9

Decrement the value of currentWeapon in your if statement, after you update the text.