Update script.js
Step 166

for loops are declared with three expressions separated by semicolons. for (a; b; c), where a is the initialization expression, b is the condition, and c is the final expression.

The initialization expression is executed only once, before the loop starts, and is often used to define and set up the loop variable. Think of it like declaring a counter to use in your loop.

Many for loops use i as the counter and start from 0, so change let x = 1; to let i = 0;.