Update script.js
Step 170

The .includes() method determines if an array contains an element and will return either true or false.

Here is an example of the .includes() syntax:

const numbersArray = [1, 2, 3, 4, 5]
const number = 3

if (numbersArray.includes(number)) {
  console.log("The number is in the array.")
}

After your for loop, add an if statement to check if the guess is in the numbers array. You can use the .includes() method to check if the array contains the guess.